requirements.txt
torch
matplotlib
gymnasium
gymnasium[atari, accept-rom-license]
gym_super_mario_bros